# Optimization using bytestring