--- a/Codec/TPTP/Import.hs
+++ b/Codec/TPTP/Import.hs
@@ -1,9 +1,11 @@
 {-# OPTIONS -Wall #-}
-module Codec.TPTP.Import(parse,parseFile
-                        ,parseWithComment,parseWithCommentFile
+module Codec.TPTP.Import(parse,parseByteString,parseFile
+                        ,parseWithComment,parseWithCommentByteString,parseWithCommentFile
                         ,Token(..)) where
 
 
+import qualified Data.ByteString.Lazy as BL
+import qualified Data.ByteString.Lazy.Builder as Builder
 import Lexer
 import Parser
 import ParserC
@@ -11,14 +13,20 @@ import Codec.TPTP.Base
 
 
 parse :: String -> [TPTP_Input]
-parse = parseTPTP . map snd . alexScanTokens
+parse = parseByteString . Builder.toLazyByteString . Builder.stringUtf8
+
+parseByteString :: BL.ByteString -> [TPTP_Input]
+parseByteString = parseTPTP . map snd . alexScanTokens
 
 parseFile :: FilePath -> IO [TPTP_Input]
-parseFile x = parse `fmap` readFile x
+parseFile x = parseByteString `fmap` BL.readFile x
 
 
 parseWithComment :: String -> [TPTP_Input_C]
-parseWithComment = parseTPTPwithComment . map snd . alexScanTokens
+parseWithComment = parseWithCommentByteString . Builder.toLazyByteString . Builder.stringUtf8
+
+parseWithCommentByteString :: BL.ByteString -> [TPTP_Input_C]
+parseWithCommentByteString = parseTPTPwithComment . map snd . alexScanTokens
 
 parseWithCommentFile :: FilePath -> IO [TPTP_Input_C]
-parseWithCommentFile x = parseWithComment `fmap` readFile x
+parseWithCommentFile x = parseWithCommentByteString `fmap` BL.readFile x