--- a/logic-TPTP.cabal
+++ b/logic-TPTP.cabal
@@ -61,6 +61,7 @@ Library
 
  build-depends:      base >=4 && < 5
                    , array
+                   , bytestring
                    , syb
                    , containers
                    , ansi-wl-pprint
@@ -95,6 +96,7 @@ Executable TestImportExportImportFile
  hs-source-dirs:    testing
  build-depends:     logic-TPTP
                   , base
+                  , bytestring
                   , ansi-wl-pprint
                   , optparse-applicative >=0.11 && <0.16
                   , pcre-light
@@ -112,6 +114,7 @@ Test-suite TestImportExportRandom
  hs-source-dirs:    testing
  build-depends:     logic-TPTP
                   , base
+                  , bytestring
                   , ansi-wl-pprint
                   , pcre-light
                   , QuickCheck