--- a/logic-TPTP.cabal
+++ b/logic-TPTP.cabal
@@ -43,7 +43,6 @@ tested-with: GHC==8.4.4
 tested-with: GHC==8.2.2
 tested-with: GHC==8.0.2
 tested-with: GHC==7.10.3
-tested-with: GHC==7.8.4
 
 source-repository head
  type: git
@@ -59,7 +58,7 @@ Flag BuildTestPrograms
 Library
  ghc-options: -Wall -O2
 
- build-depends:      base >=4 && < 5
+ build-depends:      base >=4.8 && < 5
                    , array
                    , bytestring
                    , syb